requirements.txt
streamlit
opencv-python-headless
ultralytics
torch
torchvision
torchaudio
pillow
numpy